Fix for code scanning alert no. 2: "Workflow does not contain permissions"
Fix for https://github.com/sailfishos-applications/filecase/security/code-scanning/3

To fix the issue, a permissions block is added at the root of the workflow file. This block defines the minimal permissions required for the workflow to function correctly. Here `contents: read` and `actions: write` permissions are set.

The permissions block is added right after the `name` field to apply these permissions globally to all jobs in the workflow.